Move SiteError to types folder
