+++ src/pages/patientView/clinicalInformation/PatientViewPageStore.ts
@@ -136,7 +136,7 @@ import { groupTrialMatchesById } from '../trialMatch/TrialMatchTableUtils';
 import { GeneFilterOption } from '../mutation/GeneFilterMenu';
 import TumorColumnFormatter from '../mutation/column/TumorColumnFormatter';
 import { getVariantAlleleFrequency } from 'shared/lib/MutationUtils';
-import { AppStore, SiteError } from '../../../AppStore';
+import { AppStore } from '../../../AppStore';
 import { getGeneFilterDefault } from './PatientViewPageStoreUtil';
 import { checkNonProfiledGenesExist } from '../PatientViewPageUtils';
 import autobind from 'autobind-decorator';
@@ -195,6 +195,7 @@ import { StructuralVariantFilter } from 'cbioportal-ts-api-client';
 import { IGenePanelDataByProfileIdAndSample } from 'shared/lib/isSampleProfiled';
 import { NamespaceColumnConfig } from 'shared/components/namespaceColumns/NamespaceColumnConfig';
 import { buildNamespaceColumnConfig } from 'shared/components/namespaceColumns/namespaceColumnsUtils';
+import { SiteError } from 'shared/model/appMisc';
 
 type PageMode = 'patient' | 'sample';
 type ResourceId = string;
