+++ src/shared/lib/errorFormatter.spec.ts
@@ -1,10 +1,10 @@
 import { assert } from 'chai';
-import { SiteError } from '../../AppStore';
 import {
     formatErrorLog,
     formatErrorTitle,
     formatErrorMessages,
 } from './errorFormatter';
+import { SiteError } from 'shared/model/appMisc';
 
 const RESPONSE_BODY_MESSAGE = 'RESPONSE_BODY_MESSAGE';
 const REQUEST_ERROR_MESSAGE = 'REQUEST_ERROR_MESSAGE';
