+++ src/shared/model/appMisc.ts
@@ -1,0 +1,7 @@
+export class SiteError {
+    constructor(
+        public errorObj: any,
+        public displayType: 'alert' | 'site' = 'site',
+        public title?: string
+    ) {}
+}
